recover from invalid sequence, refactor, test
So far if the sequence file contains an invalid value the couch request for the changes feed failed and we retried again and again with the same query.

Now we will reset the since parameter to nil instead and thus restart from scratch and process all documents again. CouchRest Changes should be used in a way that allows reprocessing all documents anyway. So this should not cause any problems but instead allows us to recover from the invalid seq.

I also turned the CouchRest::Changes class into a module and moved the class into CouchRest::Changes::Observer. The module has a new method that will delegate to the new class so the old interface should still be supported. Will increase the version to 0.1.0 never the less as this is likely to break any kind of monkey patching ( and it's about time for 0.1.0 ).

In addition an initial test is provided with setup scripts to run on travis.